Better handling of timestamps
DateTime instances received via the Borger.dk web service does not specify it's kind, which could cause some issues when working the date and time later on. The updated code makes the assumption that the date and time are always specified in Danish time, and initializes new EssentialTime instances accordingly.