Removed XML related logic
We don't really need all the XML related logic anymore as we're already using HtmlAgilityPack to parse the HTML.

The XML logic was primarily used to cleanup up and manipulate the HTML received from the Borger.dk API. For one, the HTML is now a lot cleaner with less errors compared to when the logic was introduced. And any manipulation should ideally not be handled by this package, but by the code interacting with this package - eg. in our Borger.dk package for Umbraco.